Reorganized the structure of the project, fixed minor mistakes
