+++ homework-11/pkg/index/index.go
@@ -3,7 +3,7 @@ package index
 import (
 	"strings"
 
-	"go-course-4/homework-11/server/pkg/crawler"
+	"go-course-4/homework-11/pkg/crawler"
 )
 
 // Index - служба инвертированного индексирования.
